+++ pulled_search.py
@@ -89,75 +89,40 @@
     Configuration file (config/search.py.TEMPLATE).  Below is the
     configuration file format for the environment setup in the program.
 
-    ###########################################################################
-    # Pulled Search General Configuration section.
-    # This section is for all options (-F, -I, -P).
-    #
     # Logger file for the storage of log entries.
-    # File name including directory path.
     log_file = "BASE_PATH/log/pulled_search.log"
 
-    ###########################################################################
-    # Pulled Search Process Configuration section.
-    # This section is for the -P option.
-    #
+    # Directory where raw data is saved to before filtering.
+    raw_archive_dir = "BASE_PATH/raw_archive"
+    # Directory where unparsable data is saved to.
+    unparsable_dir = "BASE_PATH/unparsable"
+
     # Directory where Docid Pulled Html files are located at.
-    # NOTE: Do not include the YYYY/MM as part of the path as this will be
-    #   added.
     doc_dir = ["DOC_DIR_PATH", "DOC_DIR_PATH2"]
     # Path and file name for previous processed files.
     processed_file = "BASE_PATH/processed/processed"
     # Temporary file where check_log will write to.
-    # File name including directory path.
     outfile = "BASE_PATH/tmp/checklog.out"
     # Directory path to where error and non-processed files are saved to.
     error_dir = "BASE_PATH/search_error"
     # Security enclave these files are being processed on.
     enclave = "ENCLAVE"
-    # Directory where active or archived log files to be searched are.
+    # Directory where active or archive log files to be searched are.
     log_dir = "LOG_DIR_PATH"
 
-    # These options will not need to be updated normally.
-    # Regular expression for search for html file names.
-    file_regex = "-PULLED-"
-    # Regular expression for search for recalled products.
-    pattern = "JAC.pull.subtype.*.SECURITY RECALL"
-    # Type of apache log files to checked.
-    log_type = "access_log"
-    # Mapping of commands to keywords.
-    # This is for the naming of the access logs which are not always under the
-    #   command name.
-    command = {"intelink": "eucom"}
-
-    ###########################################################################
     # Email Configuration section.
-    # These entries are for the -e option under the -P and -F options.
-    #
-    # This option is normally is used in conjunction with the rmq_2_mail.py
-    #   program.
     # Email address to rabbitmq alias for the rmq_2_mail.py program.
-    # Example: to_addr = "rabbitmq@domain.name"
     to_addr = None
     # Name of the RabbitMQ queue.
-    # Note: Subject must match exactly the RabbitMQ queue name and is
-    #   case-sensitive.
-    #     Also the subject will be CamelCased when processed.
-    # Example:  subj = "Pulledsearch"
     subj = None
 
-    ###########################################################################
     # RabbitMQ Configuration section.
-    # These entries are for the -r option under the -P and -F options.
-    #
     # Login information.
     user = "USER"
     japd = "PSWORD"
     # Address to single RabbitMQ node.
     host = "HOSTNAME"
     # List of hosts along with their ports to a multiple node RabbitMQ cluster.
-    # Format of each entry is: "IP:PORT".
-    # Example: host_list = ["hostname:5672", "hostname2:5672", "hostname:5673"]
-    # Note:  If host_list is set, it will take precedence over the host entry.
     host_list = []
     # RabbitMQ Queue name.
     queue = "QUEUENAME"
@@ -166,62 +131,18 @@
     # RabbitMQ Exchange name for each instance run.
     exchange_name = "EXCHANGE_NAME"
 
-    # NOTE: These options will not need to be updated normally.
-    # RabbitMQ listening port
-    # Default is 5672
-    port = 5672
-    # Type of exchange
-    # Names allowed:  direct, topic, fanout, headers
-    exchange_type = "direct"
-    # Is exchange durable: True|False
-    x_durable = True
-    # Are queues durable: True|False
-    q_durable = True
-    # Do queues automatically delete once message is processed:  True|False
-    auto_delete = False
-
-    ###########################################################################
-    # Pulled Search Insert Configuration section.
-    # These entries are for the -I option.
-    #
     # Directory where to monitor for new files to insert into Mongodb.
     monitor_dir = "MONITOR_DIR_PATH"
     # Regular expression for search for Insert/Mongodb file names.
     mfile_regex = "_mongo.json"
     # Directory path to where Insert/Mongodb error and non-processed files are
-    #   saved to.
+    # saved to.
     merror_dir = "BASE_PATH/mongo_error"
 
-    ###########################################################################
-    # Name of Mongo configuration file.  (Do not include the ".py" in the
-    #   name.)
-    # These entries are for the -i and -I options (mongo database).
-    #
     # Directory path to where Insert/Mongodb archived files are saved to.
-    marchive_dir = "BASE_PATH/archive"
-    # Do not change unless changing the name of the external Mongo config file.
+    marchive_dir = "BASE_PATH/mongo_archive"
+    # The config file is saved to the same location as the -d option.
     mconfig = "mongo"
-
-    ###########################################################################
-    # Log parsing section.
-    # These entries are for the -I and -P options.
-    #
-    # Warning: Do not modify this section unless you know regular expressions.
-    # NOTE: These name tags are reserved and cannot be used:
-    #   ["command", "docid", "network", "pubDate", "asOf"]
-    regex = "(?P<ip>.*?) (?P<proxyid>.*?) (?P<userid>CN=.*?) \[(?P<logTime>.*?)
-        (?= ) (?P<timeZone>.*?)\] (?P<requestid>.*?)
-        (?P<secs>.*?)/(?P<msecs>.*?) \"(?P<verb>.*?) (?P<verbUrl>.*?)
-        HTTP/(?P<httpVer>.*?)\" (?P<status>.*?) (?P<length>.*?)
-        \"(?P<referrer>.*?)\" \"(?P<userAgent>.*?)\" (?P<url>.*?)?$"
-    # These are the entries that will be parsed from the log entry and placed
-    #   into the document.
-    # Note 1: Name tags must match between regex and allowable and are
-    #   case-sensitive.
-    # Note 2: The "url" tag is hardcoded in the program to add "https://" to
-    #   the front of the url.
-    allowable = ["userid", "logTime", "verb", "status", "url"]
-
 
     Mongo configuration file format (config/mongo.py.TEMPLATE).  The
     configuration file format is for connecting to a Mongo database or
@@ -229,75 +150,49 @@
     be used to connect to a Mongo database or replica set to insert the
     results of the performance monitoring into.
 
-    # Pulled Search Insert/Mongo DB Configuration section.
-    # Update this file if using the -I option.
     user = "USER"
     japd = "PSWORD"
     # Mongo DB host information
     host = "HOST_IP"
     name = "HOSTNAME"
     # Mongo database port
-    # Default port for Mongo is 27017.
     port = 27017
     # Mongo configuration settings
-    # Only set if using a different Mongo configuration file.
     conf_file = None
     # Authentication required:  True|False
-    # Only set to False if no authentication is taking place.
     auth = True
     # Authentication database
-    # Name of database to authenticate the user in.
     auth_db = "admin"
     # Authentication mechanism
-    #   Current values allowed:  MONGODB-CR, SCRAM-SHA-1, SCRAM-SHA-256
-    #   NOTE 1:  SCRAM-SHA-256 only works for Mongodb 4.0 and better.
-    #   NOTE 2:  FIPS 140-2 environment requires SCRAM-SHA-1 or SCRAM-SHA-256.
-    #   NOTE 3:  MONGODB-CR is not suppoerted in Mongodb 4.0 and better.
     auth_mech = "SCRAM-SHA-1"
 
-    # Replica Set Mongo configuration settings.
-    # By default all settings are set to None.
-    #    None means the Mongo database is not part of a replica set.
-    #
     # Replica set name.
     # Format:  repset = "REPLICA_SET_NAME"
     repset = None
     # Replica host listing.
-    # Format:  repset_hosts = "HOST1:PORT, HOST2:PORT, [...]"
     repset_hosts = None
     # Database to authentication to.
-    # Format:  db_auth = "AUTHENTICATION_DATABASE"
     db_auth = None
 
     # Authorization Type
-    # If not set will connect to Mongo without using TLS or SSL connections.
-    # Values: TLS | SSL | None
     auth_type = None
 
     # SSL Configuration settings
-    # If not set will connect to Mongo without using SSL connections.
     # File containing the SSL certificate authority.
-    # Example: ssl_client_ca = "/opt/mongo/certs/ca.pem"
     ssl_client_ca = None
     # File containing the SSL key.
-    # Example:  ssl_client_key = "/opt/mongo/certs/client-key.pem"
     ssl_client_key = None
     # File containing the SSL certificate file.
-    # Example: ssl_client_cert = "/opt/mongo/certs/client-cert.pem"
     ssl_client_cert = None
     # Pass phrase for the SSL Client Key, if one is set.
-    # Example:  ssl_client_phrase = "Phrase"
     ssl_client_phrase = None
 
     # TLS Configuration settings
     # File containing the TLS certificate authority.
-    # Example: tls_ca_certs = "/opt/mongo/certs/ca.pem"
     tls_ca_certs = None
     # File containing the TLS Certificate and key.
-    # Example:  tls_certkey = "/opt/mongo/certs/client-key.pem"
     tls_certkey = None
     # Pass phrase for the TLS Client Key, if one is set.
-    # Example:  tls_certkey_phrase = "Phrase"
     tls_certkey_phrase = None
 
     # Name of Mongo database for data insertion
@@ -571,6 +466,63 @@ def insert_mongo(args, cfg, log, data, **kwargs):
     return status
 
 
+def filter_data(args, cfg, log, log_json):
+
+    """Function:  filter_data
+
+    Description:  Filter out non-required data entries.
+
+    Arguments:
+        (input) args -> ArgParser class instance
+        (input) cfg -> Configuration setup
+        (input) log -> Log class instance
+        (input) log_json -> Dictionary log document
+        (output) log_json -> Modified dictionary log document
+
+    """
+
+    log_json = dict(log_json)
+    log.log_info("filter_data:  Writing to raw data toarchive: %s" \
+                 % (cfg.raw_archive_dir))
+    fname = os.path.join(
+        cfg.raw_archive_dir, log_json["docid"] + "." + log_json["asOf"] +
+        ".raw")
+    gen_libs.write_file(fname=fname, mode="w", data=log_json)
+    log.log_info("filter_data:  Start filtering JSON document.")
+
+    # Loop on servers
+    for svr in log_json["servers"]:
+        ufname = os.path.join(
+            cfg.unparsable_dir, log_json["docid"] + "." + svr +
+            log_json["asOf"] + ".unparsable")
+        parsed_list = list()
+
+        # Loop on log entries for each server
+        for line in log_json["servers"][svr]:
+            parsed_line = re.match(cfg.regex, line)
+
+            # Parse the log entry
+            if parsed_line:
+                parsed_line = parsed_line.groupdict()
+
+                # Filter out non-related entries
+                if log_json["docid"] in parsed_line["url"]      \
+                   and "transformer" in parsed_line["url"]      \
+                   and "2ndReview" not in parsed_line["url"]    \
+                   and parsed_line["status"] == "200"           \
+                   and ".ic.gov" not in parsed_line["userid"]:
+                    parsed_list.append(line)
+
+            else:
+                log.log_warn("filter_data:  Unparsable data written to %s." %
+                             (ufname))
+                gen_libs.write_file(fname=uname, mode="a", data=line)
+
+        log_json["servers"][svr] = parsed_list
+
+    return log_json
+
+
 def parse_data(args, cfg, log, log_json):
 
     """Function:  parse_data
@@ -606,6 +558,7 @@ def parse_data(args, cfg, log, log_json):
         second_stage["server"] = svr
         third_stage = dict(second_stage)
 
+#        print("LENGTH_LIST", len(log_json["servers"][svr]))
         # Loop on log entries for each server
         for line in log_json["servers"][svr]:
             third_stage["entry"] = line
@@ -623,6 +576,7 @@ def parse_data(args, cfg, log, log_json):
                    and ".ic.gov" not in parsed_line["userid"]:
 
                     for entry in parsed_line:
+#                        print("ENTRY", entry)
                         if entry in cfg.allowable and entry == "url":
                             third_stage[entry] = \
                                 "https://" + parsed_line[entry]
@@ -630,7 +584,10 @@ def parse_data(args, cfg, log, log_json):
                         elif entry in cfg.allowable:
                             third_stage[entry] = parsed_line[entry]
 
-                    status = status & insert_mongo(args, cfg, log, third_stage)
+                    #status = status & insert_mongo(args, cfg, log, third_stage)
+                    #print('PROCESSED')
+                    print('PROCESSED', len(str(third_stage)))
+                    #print(third_stage)
 
             else:
                 log.log_err("parse_data:  Unable to parse log entry: %s."
@@ -695,6 +652,9 @@ def process_json(args, cfg, log, log_json):
 
     log.log_info("process_json:  Processing JSON document.")
     status = False
+
+    # Filter the raw data
+    log_json = filter_data(args, cfg, log, log_json)
 
     # Insert entries into Mongo
     if args.arg_exist("-i"):